--- test/odbc/psqlodbc/test_pg_datatype_uniqueidentifier.cpp
@@ -1,876 +1,0 @@
-#include <gtest/gtest.h>
-#include <sqlext.h>
-#include "../src/odbc_handler.h"
-#include "../src/query_generator.h"
-#include "../src/drivers.h"
-#include <cmath>
-#include <iostream>
-#include <regex>
-using std::pair;
-
-const string TABLE_NAME = "master_dbo.unique_identifier_table_odbc_test";
-const string COL1_NAME = "pk";
-const string COL2_NAME = "data";
-const string DATATYPE_NAME = "sys.uniqueidentifier";
-const string VIEW_NAME = "master_dbo.unique_identifier_view_odbc_test";
-const vector<pair<string, string>> TABLE_COLUMNS = {
-  {COL1_NAME, "INT PRIMARY KEY"},
-  {COL2_NAME, DATATYPE_NAME}
-};
-const int DATA_COLUMN = 2;
-const int BUFFER_SIZE = 256;
-const int INT_BYTES_EXPECTED = 4;
-const int GUID_BYTES_EXPECTED = 36;
-// Format for GUID: XXXXXXXX-XXXX-XXXX-XXXX-XXXXXXXXXXXX
-//                  Where X is a hexadecimal value (0-F)
-const std::regex GUID_REGEX("[\\da-fA-F]{8}[-]([0-9a-fA-F]{4}-){3}[\\da-fA-F]{12}");
-
-class PSQL_DataTypes_UniqueIdentifier : public testing::Test {
-  void SetUp() override {
-    if (!Drivers::DriverExists(ServerType::PSQL)) {
-      GTEST_SKIP() << "PSQL Driver not present: skipping all tests for this fixture.";
-    }
-
-    OdbcHandler test_setup(Drivers::GetDriver(ServerType::PSQL));
-    test_setup.ConnectAndExecQuery(DropObjectStatement("TABLE", TABLE_NAME));
-  }
-
-  void TearDown() override {
-    OdbcHandler test_teardown(Drivers::GetDriver(ServerType::PSQL));
-    test_teardown.ConnectAndExecQuery(DropObjectStatement("VIEW", VIEW_NAME));
-    test_teardown.CloseStmt();
-    test_teardown.ExecQuery(DropObjectStatement("TABLE", TABLE_NAME));
-  }
-};
-
-TEST_F(PSQL_DataTypes_UniqueIdentifier, Table_Creation) {
-  // TODO - Expected needs to be fixed.
-  const int LENGTH_EXPECTED = 255;
-  const int PRECISION_EXPECTED = 0;
-  const int SCALE_EXPECTED = 0;
-  const string NAME_EXPECTED = "unknown";
-
-  char name[BUFFER_SIZE];
-  SQLLEN length;
-  SQLLEN precision;
-  SQLLEN scale;
-
-  RETCODE rcode;
-  OdbcHandler odbcHandler(Drivers::GetDriver(ServerType::PSQL));
-
-  // Create a table with columns defined with the specific datatype being tested.
-  odbcHandler.ConnectAndExecQuery(CreateTableStatement(TABLE_NAME, TABLE_COLUMNS));
-  odbcHandler.CloseStmt();
-
-  // Select * From Table to ensure that it exists
-  odbcHandler.ExecQuery(SelectStatement(TABLE_NAME, {"*"}, vector<string>{COL1_NAME}));
-
-  // Make sure column attributes are correct
-  rcode = SQLColAttribute(odbcHandler.GetStatementHandle(),
-                          DATA_COLUMN,
-                          SQL_DESC_LENGTH, // Get the length of the column (size of char in columns)
-                          NULL,
-                          0,
-                          NULL,
-                          (SQLLEN *)&length);
-  ASSERT_EQ(rcode, SQL_SUCCESS);
-  ASSERT_EQ(length, LENGTH_EXPECTED);
-
-  rcode = SQLColAttribute(odbcHandler.GetStatementHandle(),
-                          DATA_COLUMN,
-                          SQL_DESC_PRECISION, // Get the precision of the column
-                          NULL,
-                          0,
-                          NULL,
-                          (SQLLEN *)&precision);
-  ASSERT_EQ(rcode, SQL_SUCCESS);
-  ASSERT_EQ(precision, PRECISION_EXPECTED);
-
-  rcode = SQLColAttribute(odbcHandler.GetStatementHandle(),
-                          DATA_COLUMN,
-                          SQL_DESC_SCALE, // Get the scale of the column
-                          NULL,
-                          0,
-                          NULL,
-                          (SQLLEN *)&scale);
-  ASSERT_EQ(rcode, SQL_SUCCESS);
-  ASSERT_EQ(scale, SCALE_EXPECTED);
-
-  rcode = SQLColAttribute(odbcHandler.GetStatementHandle(),
-                          DATA_COLUMN,
-                          SQL_DESC_TYPE_NAME, // Get the type name of the column
-                          name,
-                          BUFFER_SIZE,
-                          NULL,
-                          NULL);
-  ASSERT_EQ(string(name), NAME_EXPECTED);
-
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  ASSERT_EQ(rcode, SQL_NO_DATA);
-
-  odbcHandler.CloseStmt();
-  odbcHandler.ExecQuery(DropObjectStatement("TABLE", TABLE_NAME));
-}
-
-TEST_F(PSQL_DataTypes_UniqueIdentifier, Insertion_Success) {
-  int pk;
-  char data[BUFFER_SIZE];
-  SQLLEN pk_len;
-  SQLLEN data_len;
-  SQLLEN affected_rows;
-
-  RETCODE rcode;
-  OdbcHandler odbcHandler(Drivers::GetDriver(ServerType::PSQL));
-
-  const vector<string> VALID_INSERTED_VALUES = {
-    "NULL",
-    "00000000-0000-0000-0000-000000000000", // Min
-    "0E984725-C51C-4BF4-9960-E1C80E27ABA0", // Rand
-    "FFFFFFFF-FFFF-FFFF-FFFF-FFFFFFFFFFFF"  // Max
-  };
-  const int NUM_OF_INSERTS = VALID_INSERTED_VALUES.size();
-
-  const vector<tuple<int, int, SQLPOINTER, int, SQLLEN *>> BIND_COLUMNS = {
-    {1, SQL_C_LONG, &pk, 0, &pk_len},
-    {2, SQL_C_CHAR, &data, BUFFER_SIZE, &data_len}
-  };
-
-  string insert_string{};
-  string comma{};
-
-  for (int i = 0; i < NUM_OF_INSERTS; i++) {
-    const string INSERT_VALUE = VALID_INSERTED_VALUES[i] != "NULL" ? "\'" + VALID_INSERTED_VALUES[i] + "\'" : VALID_INSERTED_VALUES[i];
-    insert_string += comma + "(" + std::to_string(i) + "," + INSERT_VALUE + ")";
-    comma = ",";
-  }
-
-  odbcHandler.ConnectAndExecQuery(CreateTableStatement(TABLE_NAME, TABLE_COLUMNS));
-  odbcHandler.CloseStmt();
-
-  // Insert valid values into the table and assert affected rows
-  odbcHandler.ExecQuery(InsertStatement(TABLE_NAME, insert_string));
-
-  rcode = SQLRowCount(odbcHandler.GetStatementHandle(), &affected_rows);
-  ASSERT_EQ(rcode, SQL_SUCCESS);
-  ASSERT_EQ(affected_rows, NUM_OF_INSERTS);
-
-  odbcHandler.CloseStmt();
-
-  // Select all from the tables and assert that the following attributes of the type is correct:
-  odbcHandler.ExecQuery(SelectStatement(TABLE_NAME, {"*"}, vector<string>{COL1_NAME}));
-
-  // Make sure inserted values are correct
-  ASSERT_NO_FATAL_FAILURE(odbcHandler.BindColumns(BIND_COLUMNS));
-
-  for (int i = 0; i < NUM_OF_INSERTS; i++) {
-    rcode = SQLFetch(odbcHandler.GetStatementHandle()); // retrieve row-by-row
-    ASSERT_EQ(rcode, SQL_SUCCESS);
-    ASSERT_EQ(pk_len, INT_BYTES_EXPECTED);
-    ASSERT_EQ(pk, i);
-    if (VALID_INSERTED_VALUES[i] != "NULL") {
-      ASSERT_EQ(data_len, GUID_BYTES_EXPECTED);
-      ASSERT_EQ(data, VALID_INSERTED_VALUES[i]);
-    }
-    else {
-      ASSERT_EQ(data_len, SQL_NULL_DATA);
-    }
-  }
-
-  // Assert that there is no more data
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  ASSERT_EQ(rcode, SQL_NO_DATA);
-
-  odbcHandler.CloseStmt();
-  odbcHandler.ExecQuery(DropObjectStatement("TABLE", TABLE_NAME));
-}
-
-TEST_F(PSQL_DataTypes_UniqueIdentifier, Insertion_With_Function_Success) {
-  int pk;
-  char data[BUFFER_SIZE];
-  SQLLEN pk_len;
-  SQLLEN data_len;
-  SQLLEN affected_rows;
-
-  RETCODE rcode;
-  OdbcHandler odbcHandler(Drivers::GetDriver(ServerType::PSQL));
-
-  const vector<string> VALID_INSERTED_VALUES = {
-    "sys.NEWSEQUENTIALID()",
-    "sys.newid()"
-  };
-  const int NUM_OF_INSERTS = VALID_INSERTED_VALUES.size();
-
-  const vector<tuple<int, int, SQLPOINTER, int, SQLLEN *>> BIND_COLUMNS = {
-    {1, SQL_C_LONG, &pk, 0, &pk_len},
-    {2, SQL_C_CHAR, &data, BUFFER_SIZE, &data_len}
-  };
-
-  string insert_string{};
-  string comma{};
-
-  for (int i = 0; i < NUM_OF_INSERTS; i++) {
-    insert_string += comma + "(" + std::to_string(i) + "," + VALID_INSERTED_VALUES[i] + ")";
-    comma = ",";
-  }
-
-  odbcHandler.ConnectAndExecQuery(CreateTableStatement(TABLE_NAME, TABLE_COLUMNS));
-  odbcHandler.CloseStmt();
-
-  // Insert valid values into the table and assert affected rows
-  odbcHandler.ExecQuery(InsertStatement(TABLE_NAME, insert_string));
-
-  rcode = SQLRowCount(odbcHandler.GetStatementHandle(), &affected_rows);
-  ASSERT_EQ(rcode, SQL_SUCCESS);
-  ASSERT_EQ(affected_rows, NUM_OF_INSERTS);
-
-  odbcHandler.CloseStmt();
-
-  // Select all from the tables and assert that the following attributes of the type is correct:
-  odbcHandler.ExecQuery(SelectStatement(TABLE_NAME, {"*"}, vector<string>{COL1_NAME}));
-
-  // Make sure inserted values are correct
-  ASSERT_NO_FATAL_FAILURE(odbcHandler.BindColumns(BIND_COLUMNS));
-
-  for (int i = 0; i < NUM_OF_INSERTS; i++) {
-    rcode = SQLFetch(odbcHandler.GetStatementHandle()); // retrieve row-by-row
-    ASSERT_EQ(rcode, SQL_SUCCESS);
-    ASSERT_EQ(pk_len, INT_BYTES_EXPECTED);
-    ASSERT_EQ(pk, i);
-    ASSERT_EQ(data_len, GUID_BYTES_EXPECTED);
-    ASSERT_TRUE(std::regex_match(data, GUID_REGEX));
-  }
-
-  // Assert that there is no more data
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  ASSERT_EQ(rcode, SQL_NO_DATA);
-
-  odbcHandler.CloseStmt();
-  odbcHandler.ExecQuery(DropObjectStatement("TABLE", TABLE_NAME));
-}
-
-TEST_F(PSQL_DataTypes_UniqueIdentifier, Insertion_Fail) {
-  RETCODE rcode;
-  OdbcHandler odbcHandler(Drivers::GetDriver(ServerType::PSQL));
-
-  const vector<string> INVALID_INSERTED_VALUES = {
-    "XXXXXXXX-XXXX-XXXX-XXXX-XXXXXXXXXXXX",   // Not Hex values
-    "00000000-0000-0000-0000-00000000000",    // Too short
-    "00000000-0000-0000-0000-0000000000000",  // Too long
-    "00000000:0000:0000:0000:000000000000",   // Wrong format
-    "00000000-0000-0000-0000-00000000000X",   // Invalid Character
-    "123456789"                               // Wrong format, No casting
-  };
-  const int NUM_OF_INSERTS = INVALID_INSERTED_VALUES.size();
-
-  odbcHandler.ConnectAndExecQuery(CreateTableStatement(TABLE_NAME, TABLE_COLUMNS));
-  odbcHandler.CloseStmt();
-
-  // Attempt to insert values that are out of range and assert that they all fail
-  for (int i = 0; i < NUM_OF_INSERTS; i++) {
-    string insert_string = "(" + std::to_string(i) + "," + INVALID_INSERTED_VALUES[i] + ")";
-
-    rcode = SQLExecDirect(odbcHandler.GetStatementHandle(), (SQLCHAR *)InsertStatement(TABLE_NAME, insert_string).c_str(), SQL_NTS);
-    ASSERT_EQ(rcode, SQL_ERROR);
-  }
-
-  // Select all from the tables and assert that nothing was inserted
-  odbcHandler.ExecQuery(SelectStatement(TABLE_NAME, {"*"}, vector<string>{COL1_NAME}));
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  ASSERT_EQ(rcode, SQL_NO_DATA);
-
-  odbcHandler.ExecQuery(DropObjectStatement("TABLE", TABLE_NAME));
-}
-
-TEST_F(PSQL_DataTypes_UniqueIdentifier, Update_Success) {
-  const string PK_INSERTED = "0";
-  const string DATA_INSERTED = "01234567-1234-1234-1234-0123456789AB";
-
-  const vector <string> DATA_UPDATED_VALUES = {
-    "NULL",
-    "00000000-0000-0000-0000-000000000000", // Min
-    "0E984725-C51C-4BF4-9960-E1C80E27ABA0", // Rand
-    "FFFFFFFF-FFFF-FFFF-FFFF-FFFFFFFFFFFF"  // Max
-  };
-  const int NUM_OF_DATA = DATA_UPDATED_VALUES.size();
-
-  const string INSERT_STRING = "(" + PK_INSERTED + ",\'" + DATA_INSERTED + "\')";
-  const string UPDATE_WHERE_CLAUSE = COL1_NAME + " = " + PK_INSERTED;
-
-  const int AFFECTED_ROWS_EXPECTED = 1;
-
-  int pk;
-  char data[BUFFER_SIZE];
-  SQLLEN pk_len;
-  SQLLEN data_len;
-  SQLLEN affected_rows;
-
-  RETCODE rcode;
-  OdbcHandler odbcHandler(Drivers::GetDriver(ServerType::PSQL));
-
-  const vector<tuple<int, int, SQLPOINTER, int, SQLLEN *>> BIND_COLUMNS = {
-    {1, SQL_C_LONG, &pk, 0, &pk_len},
-    {2, SQL_C_CHAR, &data, BUFFER_SIZE, &data_len}
-  };
-
-  vector<pair<string, string>> update_col{};
-
-  for (int i = 0; i < NUM_OF_DATA; i++) {
-    const string INSERT_VALUE = DATA_UPDATED_VALUES[i] != "NULL" ? "\'" + DATA_UPDATED_VALUES[i] + "\'" : DATA_UPDATED_VALUES[i];
-    update_col.push_back(pair<string, string>(COL2_NAME, INSERT_VALUE));
-  }
-
-  odbcHandler.ConnectAndExecQuery(CreateTableStatement(TABLE_NAME, TABLE_COLUMNS));
-  odbcHandler.CloseStmt();
-
-  // Insert valid values into the table using the correct ODBC data type mapping.
-  odbcHandler.ExecQuery(InsertStatement(TABLE_NAME, INSERT_STRING));
-  odbcHandler.CloseStmt();
-
-  // Bind Columns
-  ASSERT_NO_FATAL_FAILURE(odbcHandler.BindColumns(BIND_COLUMNS));
-
-  // Assert that value is inserted properly
-  odbcHandler.ExecQuery(SelectStatement(TABLE_NAME, {"*"}, vector<string>{COL1_NAME}));
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  ASSERT_EQ(rcode, SQL_SUCCESS);
-  ASSERT_EQ(pk_len, INT_BYTES_EXPECTED);
-  ASSERT_EQ(pk, atoi(PK_INSERTED.c_str()));
-  ASSERT_EQ(data_len, GUID_BYTES_EXPECTED);
-  ASSERT_EQ(data, DATA_INSERTED);
-
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  ASSERT_EQ(rcode, SQL_NO_DATA);
-  odbcHandler.CloseStmt();
-
-  // Update value multiple times
-  for (int i = 0; i < NUM_OF_DATA; i++) {
-    odbcHandler.ExecQuery(UpdateTableStatement(TABLE_NAME, vector<pair<string, string>>{update_col[i]}, UPDATE_WHERE_CLAUSE));
-
-    rcode = SQLRowCount(odbcHandler.GetStatementHandle(), &affected_rows);
-    ASSERT_EQ(rcode, SQL_SUCCESS);
-    ASSERT_EQ(affected_rows, AFFECTED_ROWS_EXPECTED);
-
-    odbcHandler.CloseStmt();
-
-    // Assert that updated value is present
-    odbcHandler.ExecQuery(SelectStatement(TABLE_NAME, {"*"}, vector<string>{COL1_NAME}));
-    rcode = SQLFetch(odbcHandler.GetStatementHandle());
-
-    ASSERT_EQ(rcode, SQL_SUCCESS);
-    ASSERT_EQ(pk_len, INT_BYTES_EXPECTED);
-    ASSERT_EQ(pk, atoi(PK_INSERTED.c_str()));
-    if (DATA_UPDATED_VALUES[i] != "NULL") {
-      ASSERT_EQ(data_len, GUID_BYTES_EXPECTED);
-      ASSERT_EQ(data, DATA_UPDATED_VALUES[i]);
-    }
-    else {
-      ASSERT_EQ(data_len, SQL_NULL_DATA);
-    }
-
-    rcode = SQLFetch(odbcHandler.GetStatementHandle());
-    ASSERT_EQ(rcode, SQL_NO_DATA);
-    odbcHandler.CloseStmt();
-  }
-
-  odbcHandler.ExecQuery(DropObjectStatement("TABLE", TABLE_NAME));
-}
-
-TEST_F(PSQL_DataTypes_UniqueIdentifier, Update_Fail) {
-  const string PK_INSERTED = "0";
-  const string DATA_INSERTED = "01234567-1234-1234-1234-0123456789AB";
-  const string DATA_UPDATED_VALUE = "01234567-1234-1234-1234-0123456wrong";
-
-  const string INSERT_STRING = "(" + PK_INSERTED + ",\'" + DATA_INSERTED + "\')";
-  const string UPDATE_WHERE_CLAUSE = COL1_NAME + " = " + PK_INSERTED;
-
-  int pk;
-  char data[BUFFER_SIZE];
-  SQLLEN pk_len;
-  SQLLEN data_len;
-
-  RETCODE rcode;
-  OdbcHandler odbcHandler(Drivers::GetDriver(ServerType::PSQL));
-
-  const vector<tuple<int, int, SQLPOINTER, int, SQLLEN *>> BIND_COLUMNS = {
-    {1, SQL_C_LONG, &pk, 0, &pk_len},
-    {2, SQL_C_CHAR, &data, BUFFER_SIZE, &data_len}
-  };
-
-  const vector<pair<string, string>> UPDATE_COL = {
-    {COL2_NAME, DATA_UPDATED_VALUE}
-  };
-
-  odbcHandler.ConnectAndExecQuery(CreateTableStatement(TABLE_NAME, TABLE_COLUMNS));
-  odbcHandler.CloseStmt();
-
-  // Insert valid values into the table using the correct ODBC data type mapping.
-  odbcHandler.ExecQuery(InsertStatement(TABLE_NAME, INSERT_STRING));
-  odbcHandler.CloseStmt();
-
-  // Bind Columns
-  ASSERT_NO_FATAL_FAILURE(odbcHandler.BindColumns(BIND_COLUMNS));
-
-  // Assert that value is inserted properly
-  odbcHandler.ExecQuery(SelectStatement(TABLE_NAME, {"*"}, vector<string>{COL1_NAME}));
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  ASSERT_EQ(rcode, SQL_SUCCESS);
-  ASSERT_EQ(pk_len, INT_BYTES_EXPECTED);
-  ASSERT_EQ(pk, atoi(PK_INSERTED.c_str()));
-  ASSERT_EQ(data_len, GUID_BYTES_EXPECTED);
-  ASSERT_EQ(data, DATA_INSERTED);
-
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  ASSERT_EQ(rcode, SQL_NO_DATA);
-  odbcHandler.CloseStmt();
-
-  // Update value and assert an error is present
-  rcode = SQLExecDirect(odbcHandler.GetStatementHandle(), (SQLCHAR *)UpdateTableStatement(TABLE_NAME, UPDATE_COL, UPDATE_WHERE_CLAUSE).c_str(), SQL_NTS);
-  ASSERT_EQ(rcode, SQL_ERROR);
-  odbcHandler.CloseStmt();
-
-  // Assert that no values changed
-  odbcHandler.ExecQuery(SelectStatement(TABLE_NAME, {"*"}, vector<string>{COL1_NAME}));
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-
-  ASSERT_EQ(rcode, SQL_SUCCESS);
-  ASSERT_EQ(pk_len, INT_BYTES_EXPECTED);
-  ASSERT_EQ(pk, atoi(PK_INSERTED.c_str()));
-  ASSERT_EQ(data_len, GUID_BYTES_EXPECTED);
-  ASSERT_EQ(data, DATA_INSERTED);
-
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  ASSERT_EQ(rcode, SQL_NO_DATA);
-
-  odbcHandler.CloseStmt();
-  odbcHandler.ExecQuery(DropObjectStatement("TABLE", TABLE_NAME));
-}
-
-// Explicit casting is used, ie OPERATOR(sys.=)
-TEST_F(PSQL_DataTypes_UniqueIdentifier, Comparison_Operators) {
-  const vector<pair<string, string>> TABLE_COLUMNS = {
-    {COL1_NAME, DATATYPE_NAME + " PRIMARY KEY"},
-    {COL2_NAME, DATATYPE_NAME}
-  };
-
-  RETCODE rcode;
-  OdbcHandler odbcHandler(Drivers::GetDriver(ServerType::PSQL));
-  SQLLEN affected_rows;
-  const int BYTES_EXPECTED = 1;
-
-  vector<string> INSERTED_PK = {
-    "00000000-0000-0000-0000-000000000000", // Min
-    "FFFFFFFF-FFFF-FFFF-FFFF-FFFFFFFFFFFF", // Max
-    "0E984725-C51C-4BF4-9960-E1C80E27ABA0", // Rand
-    "364FC78D-F4B6-4B33-8453-C2EDF32E8075"  // Rand
-  };
-
-  vector<string> INSERTED_DATA = {
-    "00000000-0000-0000-0000-000000000000", // Min
-    "FFFFFFFF-FFFF-FFFF-FFFF-FFFFFFFFFFFF", // Max
-    "90E044B1-3584-4B38-983A-795593AEBA3B", // Rand
-    "69F07EFC-B7A5-4621-BDB8-EFD24B3E239A"  // Rand
-  };
-  const int NUM_OF_DATA = INSERTED_DATA.size();
-
-  vector<string> OPERATIONS_QUERY = {
-    COL1_NAME + " OPERATOR(sys.=) " + COL2_NAME,
-    COL1_NAME + " OPERATOR(sys.<>) " + COL2_NAME,
-    COL1_NAME + " OPERATOR(sys.<) " + COL2_NAME,
-    COL1_NAME + " OPERATOR(sys.<=) " + COL2_NAME,
-    COL1_NAME + " OPERATOR(sys.>) " + COL2_NAME,
-    COL1_NAME + " OPERATOR(sys.>=) " + COL2_NAME
-  };
-  const int NUM_OF_OPERATIONS = OPERATIONS_QUERY.size();
-
-  // initialization of expected_results
-  vector<vector<char>> expected_results = {};
-  for (int i = 0; i < NUM_OF_DATA; i++) {
-    expected_results.push_back({});
-    for (auto & c: INSERTED_PK[i]) c = toupper(c);
-    const char* comp_1 = INSERTED_PK[i].c_str();
-    for (auto & c: INSERTED_DATA[i]) c = toupper(c);
-    const char* comp_2 = INSERTED_DATA[i].c_str();
-    
-    expected_results[i].push_back(strcmp(comp_1, comp_2) == 0 ? '1' : '0');
-    expected_results[i].push_back(strcmp(comp_1, comp_2) != 0 ? '1' : '0');
-    expected_results[i].push_back(strcmp(comp_1, comp_2) < 0 ? '1' : '0');
-    expected_results[i].push_back(strcmp(comp_1, comp_2) <= 0 ? '1' : '0');
-    expected_results[i].push_back(strcmp(comp_1, comp_2) > 0 ? '1' : '0');
-    expected_results[i].push_back(strcmp(comp_1, comp_2) >= 0 ? '1' : '0');
-  }
-
-  char col_results[NUM_OF_OPERATIONS];
-  SQLLEN col_len[NUM_OF_OPERATIONS];
-  vector<tuple<int, int, SQLPOINTER, int, SQLLEN *>> BIND_COLUMNS = {};
-
-  // initialization for bind_columns
-  for (int i = 0; i < NUM_OF_OPERATIONS; i++) {
-    tuple<int, int, SQLPOINTER, int, SQLLEN *> tuple_to_insert(i + 1, SQL_C_CHAR, (SQLPOINTER)&col_results[i], BUFFER_SIZE, &col_len[i]);
-    BIND_COLUMNS.push_back(tuple_to_insert);
-  }
-
-  string insert_string{};
-  string comma{};
-
-  // insert_string initialization
-  for (int i = 0; i < NUM_OF_DATA; i++) {
-    insert_string += comma + "(\'" + INSERTED_PK[i] + "\',\'" + INSERTED_DATA[i] + "\')";
-    comma = ",";
-  }
-
-  // Create table
-  odbcHandler.ConnectAndExecQuery(CreateTableStatement(TABLE_NAME, TABLE_COLUMNS));
-  odbcHandler.CloseStmt();
-
-  // Insert valid values into the table and assert affected rows
-  odbcHandler.ExecQuery(InsertStatement(TABLE_NAME, insert_string));
-
-  rcode = SQLRowCount(odbcHandler.GetStatementHandle(), &affected_rows);
-  ASSERT_EQ(rcode, SQL_SUCCESS);
-  ASSERT_EQ(affected_rows, NUM_OF_DATA);
-
-  // Make sure inserted values are correct and operations
-  ASSERT_NO_FATAL_FAILURE(odbcHandler.BindColumns(BIND_COLUMNS));
-
-  for (int i = 0; i < NUM_OF_DATA; i++) {
-    odbcHandler.CloseStmt();
-    odbcHandler.ExecQuery(SelectStatement(TABLE_NAME, OPERATIONS_QUERY, vector<string>{}, COL1_NAME + " OPERATOR(sys.=)\'" + INSERTED_PK[i] + "\'"));
-    ASSERT_NO_FATAL_FAILURE(odbcHandler.BindColumns(BIND_COLUMNS));
-
-    rcode = SQLFetch(odbcHandler.GetStatementHandle());
-    ASSERT_EQ(rcode, SQL_SUCCESS);
-    
-    for (int j = 0; j < NUM_OF_OPERATIONS; j++) {
-      ASSERT_EQ(col_len[j], BYTES_EXPECTED);
-      ASSERT_EQ(col_results[j], expected_results[i][j]);
-    }
-  }
-
-  // Assert that there is no more data
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  ASSERT_EQ(rcode, SQL_NO_DATA);
-
-  odbcHandler.CloseStmt();
-  odbcHandler.ExecQuery(DropObjectStatement("TABLE", TABLE_NAME));
-}
-
-TEST_F(PSQL_DataTypes_UniqueIdentifier, View_Creation) {
-  const string VIEW_QUERY = "SELECT * FROM " + TABLE_NAME;
-
-  int pk;
-  char data[BUFFER_SIZE];
-  SQLLEN pk_len;
-  SQLLEN data_len;
-  SQLLEN affected_rows;
-
-  RETCODE rcode;
-  OdbcHandler odbcHandler(Drivers::GetDriver(ServerType::PSQL));
-
-  const vector<string> VALID_INSERTED_VALUES = {
-    "NULL",
-    "00000000-0000-0000-0000-000000000000", // Min
-    "0E984725-C51C-4BF4-9960-E1C80E27ABA0", // Rand
-    "FFFFFFFF-FFFF-FFFF-FFFF-FFFFFFFFFFFF"  // Max
-  };
-  const int NUM_OF_INSERTS = VALID_INSERTED_VALUES.size();
-
-  const vector<tuple<int, int, SQLPOINTER, int, SQLLEN *>> BIND_COLUMNS = {
-    {1, SQL_C_LONG, &pk, 0, &pk_len},
-    {2, SQL_C_CHAR, &data, BUFFER_SIZE, &data_len}
-  };
-
-  string insert_string{};
-  string comma{};
-
-  for (int i = 0; i < NUM_OF_INSERTS; i++) {
-    const string INSERT_VALUE = VALID_INSERTED_VALUES[i] != "NULL" ? "\'" + VALID_INSERTED_VALUES[i] + "\'" : VALID_INSERTED_VALUES[i];
-    insert_string += comma + "(" + std::to_string(i) + "," + INSERT_VALUE + ")";
-    comma = ",";
-  }
-
-  // Create Table
-  odbcHandler.ConnectAndExecQuery(CreateTableStatement(TABLE_NAME, TABLE_COLUMNS));
-  odbcHandler.CloseStmt();
-
-  // Insert valid values into the table and assert affected rows
-  odbcHandler.ExecQuery(InsertStatement(TABLE_NAME, insert_string));
-
-  rcode = SQLRowCount(odbcHandler.GetStatementHandle(), &affected_rows);
-  ASSERT_EQ(rcode, SQL_SUCCESS);
-  ASSERT_EQ(affected_rows, NUM_OF_INSERTS);
-
-  odbcHandler.CloseStmt();
-
-  // Create view
-  odbcHandler.ExecQuery(CreateViewStatement(VIEW_NAME, VIEW_QUERY));
-  odbcHandler.CloseStmt();
-
-  // Select all from the tables and assert that the following attributes of the type is correct:
-  odbcHandler.ExecQuery(SelectStatement(VIEW_NAME, {"*"}, vector<string>{COL1_NAME}));
-
-  // Make sure inserted values are correct
-  ASSERT_NO_FATAL_FAILURE(odbcHandler.BindColumns(BIND_COLUMNS));
-
-  for (int i = 0; i < NUM_OF_INSERTS; i++) {
-
-    rcode = SQLFetch(odbcHandler.GetStatementHandle()); // retrieve row-by-row
-    ASSERT_EQ(rcode, SQL_SUCCESS);
-    ASSERT_EQ(pk_len, INT_BYTES_EXPECTED);
-    ASSERT_EQ(pk, i);
-
-    if (VALID_INSERTED_VALUES[i] != "NULL") {
-      ASSERT_EQ(data_len, GUID_BYTES_EXPECTED);
-      ASSERT_EQ(data, VALID_INSERTED_VALUES[i]);
-    }
-    else {
-      ASSERT_EQ(data_len, SQL_NULL_DATA); 
-    } 
-  }
-
-  // Assert that there is no more data
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  ASSERT_EQ(rcode, SQL_NO_DATA);
-
-  odbcHandler.CloseStmt();
-  odbcHandler.ExecQuery(DropObjectStatement("VIEW", VIEW_NAME));
-
-  odbcHandler.CloseStmt();
-  odbcHandler.ExecQuery(DropObjectStatement("TABLE", TABLE_NAME));
-}
-
-TEST_F(PSQL_DataTypes_UniqueIdentifier, Table_Unique_Constraints) {
-  const vector<pair<string, string>> TABLE_COLUMNS = {
-    {COL1_NAME, "INT PRIMARY KEY"},
-    {COL2_NAME, DATATYPE_NAME + " UNIQUE"}
-  };
-  const string UNIQUE_COLUMN_NAME = COL2_NAME;
-
-  int pk;
-  char data[BUFFER_SIZE];
-  SQLLEN pk_len;
-  SQLLEN data_len;
-  SQLLEN affected_rows;
-
-  RETCODE rcode;
-  OdbcHandler odbcHandler(Drivers::GetDriver(ServerType::PSQL));
-
-  const vector<string> VALID_INSERTED_VALUES = {
-    "00000000-0000-0000-0000-000000000000", // Min
-    "0E984725-C51C-4BF4-9960-E1C80E27ABA0", // Rand
-    "FFFFFFFF-FFFF-FFFF-FFFF-FFFFFFFFFFFF"  // Max
-  };
-  const int NUM_OF_INSERTS = VALID_INSERTED_VALUES.size();
-
-  const vector<tuple<int, int, SQLPOINTER, int, SQLLEN *>> BIND_COLUMNS = {
-    {1, SQL_C_LONG, &pk, 0, &pk_len},
-    {2, SQL_C_CHAR, &data, BUFFER_SIZE, &data_len}
-  };
-
-  string insert_string{};
-  string comma{};
-
-  for (int i = 0; i < NUM_OF_INSERTS; i++) {
-    insert_string += comma + "(" + std::to_string(i) + ",\'" + VALID_INSERTED_VALUES[i] + "\')";
-    comma = ",";
-  }
-
-  odbcHandler.ConnectAndExecQuery(CreateTableStatement(TABLE_NAME, TABLE_COLUMNS));
-  odbcHandler.CloseStmt();
-
-  // Check if unique constraint still matches after creation
-  char column_name[BUFFER_SIZE];
-
-  vector<tuple<int, int, SQLPOINTER, int>> table_BIND_COLUMNS = {
-      {1, SQL_C_CHAR, column_name, BUFFER_SIZE},
-  };
-  ASSERT_NO_FATAL_FAILURE(odbcHandler.BindColumns(table_BIND_COLUMNS));
-
-  const string UNIQUE_KEY_QUERY =
-    "SELECT C.COLUMN_NAME FROM "
-    "INFORMATION_SCHEMA.TABLE_CONSTRAINTS T "
-    "JOIN INFORMATION_SCHEMA.CONSTRAINT_COLUMN_USAGE C "
-    "ON C.CONSTRAINT_NAME=T.CONSTRAINT_NAME "
-    "WHERE "
-    "C.TABLE_NAME='" + TABLE_NAME.substr(TABLE_NAME.find('.') + 1, TABLE_NAME.length()) + "' "
-    "AND T.CONSTRAINT_TYPE='UNIQUE'";
-  odbcHandler.ExecQuery(UNIQUE_KEY_QUERY);
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  ASSERT_EQ(rcode, SQL_SUCCESS);
-  ASSERT_EQ(string(column_name), UNIQUE_COLUMN_NAME);
-
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  EXPECT_EQ(rcode, SQL_NO_DATA);
-
-  odbcHandler.CloseStmt();
-
-  // Insert valid values into the table and assert affected rows
-  odbcHandler.ExecQuery(InsertStatement(TABLE_NAME, insert_string));
-
-  rcode = SQLRowCount(odbcHandler.GetStatementHandle(), &affected_rows);
-  ASSERT_EQ(rcode, SQL_SUCCESS);
-  ASSERT_EQ(affected_rows, NUM_OF_INSERTS);
-
-  odbcHandler.CloseStmt();
-
-  // Select all from the tables and assert that the following attributes of the type is correct:
-  odbcHandler.ExecQuery(SelectStatement(TABLE_NAME, {"*"}, vector<string>{COL1_NAME}));
-
-  // Make sure inserted values are correct
-  ASSERT_NO_FATAL_FAILURE(odbcHandler.BindColumns(BIND_COLUMNS));
-
-  for (int i = 0; i < NUM_OF_INSERTS; i++) {
-    rcode = SQLFetch(odbcHandler.GetStatementHandle()); // retrieve row-by-row
-    ASSERT_EQ(rcode, SQL_SUCCESS);
-    ASSERT_EQ(pk_len, INT_BYTES_EXPECTED);
-    ASSERT_EQ(pk, i);
-    ASSERT_EQ(data_len, GUID_BYTES_EXPECTED);
-    ASSERT_EQ(data, VALID_INSERTED_VALUES[i]);
-  }
-
-  // Assert that there is no more data
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  ASSERT_EQ(rcode, SQL_NO_DATA);
-
-  odbcHandler.CloseStmt();
-
-  // Attempt to insert values that violates unique constraint and assert that they all fail
-  for (int i = 0; i < 2 * NUM_OF_INSERTS; i++) {
-    string insert_string = "(" + std::to_string(i) + ",\'" + VALID_INSERTED_VALUES[i % NUM_OF_INSERTS] + "\')";
-
-    rcode = SQLExecDirect(odbcHandler.GetStatementHandle(), (SQLCHAR *)InsertStatement(TABLE_NAME, insert_string).c_str(), SQL_NTS);
-    ASSERT_EQ(rcode, SQL_ERROR);
-  }
-
-  odbcHandler.ExecQuery(DropObjectStatement("TABLE", TABLE_NAME));
-}
-
-TEST_F(PSQL_DataTypes_UniqueIdentifier, Table_Composite_Keys) {
-  const vector<pair<string, string>> TABLE_COLUMNS = {
-    {COL1_NAME, "INT"},
-    {COL2_NAME, DATATYPE_NAME}
-  };
-  const string PKTABLE_NAME = TABLE_NAME.substr(TABLE_NAME.find('.') + 1, TABLE_NAME.length());
-  const string SCHEMA_NAME = TABLE_NAME.substr(0, TABLE_NAME.find('.'));
-
-  const vector<string> PK_COLUMNS = {
-    COL1_NAME,
-    COL2_NAME
-  };
-
-  string table_constraints{"PRIMARY KEY ("};
-  string comma{};
-  for (int i = 0; i < PK_COLUMNS.size(); i++) {
-    table_constraints += comma + PK_COLUMNS[i];
-    comma = ",";
-  }
-  table_constraints += ")";
-
-  int pk;
-  char data[BUFFER_SIZE];
-  SQLLEN pk_len;
-  SQLLEN data_len;
-  SQLLEN affected_rows;
-
-  RETCODE rcode;
-  OdbcHandler odbcHandler(Drivers::GetDriver(ServerType::PSQL));
-
-  vector<string> VALID_INSERTED_VALUES = {
-    "00000000-0000-0000-0000-000000000000", // Min
-    "0E984725-C51C-4BF4-9960-E1C80E27ABA0", // Rand
-    "FFFFFFFF-FFFF-FFFF-FFFF-FFFFFFFFFFFF"  // Max
-  };
-  const int NUM_OF_INSERTS = VALID_INSERTED_VALUES.size();
-
-  const vector<tuple<int, int, SQLPOINTER, int, SQLLEN *>> BIND_COLUMNS = {
-    {1, SQL_C_LONG, &pk, 0, &pk_len},
-    {2, SQL_C_CHAR, &data, BUFFER_SIZE, &data_len}
-  };
-
-  string insert_string{};
-  comma = "";
-
-  for (int i = 0; i < NUM_OF_INSERTS; i++) {
-    insert_string += comma + "(" + std::to_string(i) + ",\'" + VALID_INSERTED_VALUES[i] + "\')";
-    comma = ",";
-  }
-
-  odbcHandler.ConnectAndExecQuery(CreateTableStatement(TABLE_NAME, TABLE_COLUMNS, table_constraints));
-  odbcHandler.CloseStmt();
-
-  // Check if composite key still matches after creation
-  char table_name[BUFFER_SIZE];
-  char column_name[BUFFER_SIZE];
-  int key_sq{};
-  char pk_name[BUFFER_SIZE];
-
-  vector<tuple<int, int, SQLPOINTER, int>> constraints_BIND_COLUMNS = {
-    {3, SQL_C_CHAR, table_name, BUFFER_SIZE},
-    {4, SQL_C_CHAR, column_name, BUFFER_SIZE},
-    {5, SQL_C_ULONG, &key_sq, BUFFER_SIZE},
-    {6, SQL_C_CHAR, pk_name, BUFFER_SIZE}
-  };
-  ASSERT_NO_FATAL_FAILURE(odbcHandler.BindColumns(constraints_BIND_COLUMNS));
-
-  rcode = SQLPrimaryKeys(odbcHandler.GetStatementHandle(), NULL, 0, (SQLCHAR *)SCHEMA_NAME.c_str(), SQL_NTS, (SQLCHAR *)PKTABLE_NAME.c_str(), SQL_NTS);
-  ASSERT_EQ(rcode, SQL_SUCCESS);
-
-  int curr_sq{0};
-  for (auto columnName : PK_COLUMNS) {
-    ++curr_sq;
-    rcode = SQLFetch(odbcHandler.GetStatementHandle());
-    ASSERT_EQ(rcode, SQL_SUCCESS);
-
-    ASSERT_EQ(string(table_name), PKTABLE_NAME);
-    ASSERT_EQ(string(column_name), columnName);
-    ASSERT_EQ(key_sq, curr_sq);
-  }
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  ASSERT_EQ(rcode, SQL_NO_DATA);
-  odbcHandler.CloseStmt();
-
-  // Insert valid values into the table and assert affected rows
-  odbcHandler.ExecQuery(InsertStatement(TABLE_NAME, insert_string));
-
-  rcode = SQLRowCount(odbcHandler.GetStatementHandle(), &affected_rows);
-  ASSERT_EQ(rcode, SQL_SUCCESS);
-  ASSERT_EQ(affected_rows, NUM_OF_INSERTS);
-
-  odbcHandler.CloseStmt();
-
-  // Select all from the tables and assert that the following attributes of the type is correct:
-  odbcHandler.ExecQuery(SelectStatement(TABLE_NAME, {"*"}, vector<string>{COL1_NAME}));
-
-  // Make sure inserted values are correct
-  ASSERT_NO_FATAL_FAILURE(odbcHandler.BindColumns(BIND_COLUMNS));
-
-  for (int i = 0; i < NUM_OF_INSERTS; i++) {
-    rcode = SQLFetch(odbcHandler.GetStatementHandle()); // retrieve row-by-row
-    ASSERT_EQ(rcode, SQL_SUCCESS);
-    ASSERT_EQ(pk_len, INT_BYTES_EXPECTED);
-    ASSERT_EQ(pk, i);
-    if (VALID_INSERTED_VALUES[i] != "NULL") {
-      ASSERT_EQ(data_len, GUID_BYTES_EXPECTED);
-      ASSERT_EQ(data, VALID_INSERTED_VALUES[i]);
-    }
-    else {
-      ASSERT_EQ(data_len, SQL_NULL_DATA);
-    }
-  }
-
-  // Assert that there is no more data
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  ASSERT_EQ(rcode, SQL_NO_DATA);
-
-  odbcHandler.CloseStmt();
-
-  // Attempt to insert values that violates composite constraint and assert that they all fail
-  for (int i = 0; i < NUM_OF_INSERTS * 2; i++) {
-    insert_string += comma + "(" + std::to_string(i) + ",\'" + VALID_INSERTED_VALUES[i % NUM_OF_INSERTS] + "\')";
-    comma = ",";
-  }
-
-  rcode = SQLExecDirect(odbcHandler.GetStatementHandle(), (SQLCHAR *)InsertStatement(TABLE_NAME, insert_string).c_str(), SQL_NTS);
-  ASSERT_EQ(rcode, SQL_ERROR);
-
-  odbcHandler.ExecQuery(DropObjectStatement("TABLE", TABLE_NAME));
-}
